Abstract ParameterValueEditors
