+++ src/Umbraco.Core/PropertyEditors/ParameterEditors/MultipleContentPickerParameterEditor.cs
@@ -1,10 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using Microsoft.Extensions.Logging;
-using Umbraco.Cms.Core.Hosting;
-using Umbraco.Cms.Core.IO;
+﻿using Umbraco.Cms.Core.IO;
 using Umbraco.Cms.Core.Models;
-using Umbraco.Cms.Core.Models.Editors;
 using Umbraco.Cms.Core.Serialization;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Strings;
@@ -36,52 +31,14 @@ namespace Umbraco.Cms.Core.PropertyEditors.ParameterEditors
 
         protected override IDataValueEditor CreateValueEditor() => DataValueEditorFactory.Create<MultipleContentPickerParameterEditor.MultipleContentPickerParamateterValueEditor>(Attribute);
 
-        internal class MultipleContentPickerParamateterValueEditor : DataValueEditor, IDataValueReference
+        internal class MultipleContentPickerParamateterValueEditor : MultiplePickerParamateterValueEditorBase
         {
-            private readonly IEntityService _entityService;
-
-            public MultipleContentPickerParamateterValueEditor(
-                ILocalizedTextService localizedTextService,
-                IShortStringHelper shortStringHelper,
-                IJsonSerializer jsonSerializer,
-                IIOHelper ioHelper,
-                DataEditorAttribute attribute,
-                IEntityService entityService)
-                : base(localizedTextService, shortStringHelper, jsonSerializer, ioHelper, attribute)
+            public MultipleContentPickerParamateterValueEditor(ILocalizedTextService localizedTextService, IShortStringHelper shortStringHelper, IJsonSerializer jsonSerializer, IIOHelper ioHelper, DataEditorAttribute attribute, IEntityService entityService) : base(localizedTextService, shortStringHelper, jsonSerializer, ioHelper, attribute, entityService)
             {
-                _entityService = entityService;
             }
 
-            public IEnumerable<UmbracoEntityReference> GetReferences(object value)
-            {
-                var asString = value is string str ? str : value?.ToString();
-
-                if (string.IsNullOrEmpty(asString))
-                {
-                    yield break;
-                }
-
-                foreach (var udiStr in asString.Split(','))
-                {
-                    if (UdiParser.TryParse(udiStr, out Udi udi))
-                    {
-                        yield return new UmbracoEntityReference(udi);
-                    }
-
-                    // this is needed to support the legacy case when the multiple media picker parameter editor stores ints not udis
-                    if (int.TryParse(udiStr, out var id))
-                    {
-                        Attempt<Guid> guidAttempt = _entityService.GetKey(id, UmbracoObjectTypes.Document);
-                        Guid guid = guidAttempt.Success ? guidAttempt.Result : Guid.Empty;
-
-                        if (guid != Guid.Empty)
-                        {
-                            yield return new UmbracoEntityReference(new GuidUdi(Constants.UdiEntityType.Media, guid));
-                        }
-
-                    }
-                }
-            }
+            public override string UdiEntityType { get; } = Constants.UdiEntityType.Document;
+            public override UmbracoObjectTypes UmbracoObjectType { get; } = UmbracoObjectTypes.Document;
         }
     }
 }
