+++ src/Umbraco.Core/PropertyEditors/ParameterEditors/MultipleMediaPickerParameterEditor.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Reflection.Metadata;
 using Umbraco.Cms.Core.IO;
 using Umbraco.Cms.Core.Models;
 using Umbraco.Cms.Core.Models.Editors;
@@ -32,52 +33,14 @@ namespace Umbraco.Cms.Core.PropertyEditors.ParameterEditors
 
         protected override IDataValueEditor CreateValueEditor() => DataValueEditorFactory.Create<MultipleMediaPickerPropertyValueEditor>(Attribute);
 
-        internal class MultipleMediaPickerPropertyValueEditor : DataValueEditor, IDataValueReference
+        internal class MultipleMediaPickerPropertyValueEditor : MultiplePickerParamateterValueEditorBase
         {
-            private readonly IEntityService _entityService;
-
-            public MultipleMediaPickerPropertyValueEditor(
-                ILocalizedTextService localizedTextService,
-                IShortStringHelper shortStringHelper,
-                IJsonSerializer jsonSerializer,
-                IIOHelper ioHelper,
-                DataEditorAttribute attribute,
-                IEntityService entityService)
-                : base(localizedTextService, shortStringHelper, jsonSerializer, ioHelper, attribute)
+            public MultipleMediaPickerPropertyValueEditor(ILocalizedTextService localizedTextService, IShortStringHelper shortStringHelper, IJsonSerializer jsonSerializer, IIOHelper ioHelper, DataEditorAttribute attribute, IEntityService entityService) : base(localizedTextService, shortStringHelper, jsonSerializer, ioHelper, attribute, entityService)
             {
-                _entityService = entityService;
             }
 
-            public IEnumerable<UmbracoEntityReference> GetReferences(object value)
-            {
-                var asString = value is string str ? str : value?.ToString();
-
-                if (string.IsNullOrEmpty(asString))
-                {
-                    yield break;
-                }
-
-                foreach (var udiStr in asString.Split(','))
-                {
-                    if (UdiParser.TryParse(udiStr, out Udi udi))
-                    {
-                        yield return new UmbracoEntityReference(udi);
-                    }
-
-                    // this is needed to support the legacy case when the multiple media picker parameter editor stores ints not udis
-                    if (int.TryParse(udiStr, out var id))
-                    {
-                        Attempt<Guid> guidAttempt = _entityService.GetKey(id, UmbracoObjectTypes.Media);
-                        Guid guid = guidAttempt.Success ? guidAttempt.Result : Guid.Empty;
-
-                        if (guid != Guid.Empty)
-                        {
-                            yield return new UmbracoEntityReference(new GuidUdi(Constants.UdiEntityType.Media, guid));
-                        }
-
-                    }
-                }
-            }
+            public override string UdiEntityType { get; } = Constants.UdiEntityType.Media;
+            public override UmbracoObjectTypes UmbracoObjectType { get; } = UmbracoObjectTypes.Media;
         }
     }
 }
